Parses wo crash

+++ Assets/Scripts/WdlEngine/Token.cs
@@ -13,7 +13,7 @@ namespace WdlEngine
 
         public string StringValue => Value.ToString();
         public float RealValue => (float)Value;
-        public float IntValue => (int)Value;
+        public int IntValue => (int)Value;
         public float NumericValue => Type == TokenType.Integer ? IntValue : RealValue;
 
         public Token(TokenType type, object value)
